pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-lektor"
version = "0.1.0"
description = "MCP server for professional German-language proofreading of Word documents"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "python-docx>=1.1.0",
    "lxml>=5.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
